fix(ingestion): Tableau test case execution fix (#6005)
